disable debug info for now
This will be re-enabled at some point, but currently it's not worth it
because

a) it increases compile time significantly, and
b) CONFIG_DEBUG_INFO_BTF causes the build to fail

Long compile times aren't an issue under normal circumstances but the
goal right now is to actually test the kernel, so a 2-3 hour build
time is not practical if the kernel needs to be built multiple times
in a day.

The issue with BTF is easily fixable but will require getting the source
from upstream rather than Debian, which means cherry-picking Debian's
patches, plus a new version of pahole. This will take time and, again,
the goal right now is to test the kernel, not fiddle with it.